requirements.txt
requests>=2.25.0
python-dotenv>=0.15.0
click>=8.0.0
tqdm>=4.62.0
